backend/requirements.txt
fastapi>=0.111
uvicorn[standard]>=0.29
httpx>=0.27
numpy>=1.26
scikit-learn>=1.4
shapely>=2.0
pyproj>=3.6
python-dotenv>=1.0
pytest>=8.0
# Optional — uncomment for large indexes (10 000+ cities)
# faiss-cpu>=1.8
